Refactor selection handling: move utility functions to polygonUtils.js and update imports
