+++ html/scripts/polygonUtils.js
@@ -1,0 +1,35 @@
+import { stepLength } from './settings.js';
+
+export function getPathPoints(path, useSimplified = false) {
+    const points = [];
+    const dAttribute = useSimplified ? 'data-simplified-d' : 'd';
+    const pathData = path.getAttribute(dAttribute);
+
+    if (!pathData) {
+        console.warn(`Path ${path.id} does not have a ${dAttribute} attribute.`);
+        return { points, svg: path.ownerSVGElement };
+    }
+
+    const pathLength = path.getTotalLength();
+    const svg = path.ownerSVGElement;
+    for (let i = 0; i <= pathLength; i += stepLength) {
+        const point = path.getPointAtLength(i);
+        if (!points.some(p => p.x === point.x && p.y === point.y)) {
+            points.push(point);
+        }
+    }
+    return { points, svg };
+}
+
+export function isPointInPolygon(point, polygon) {
+    let inside = false;
+    for (let i = 0, j = polygon.length - 1; i < polygon.length; j = i++) {
+        const xi = polygon[i].x, yi = polygon[i].y;
+        const xj = polygon[j].x, yj = polygon[j].y;
+
+        const intersect = ((yi > point.y) !== (yj > point.y))
+            && (point.x < (xj - xi) * (point.y - yi) / (yj - yi) + xi);
+        if (intersect) inside = !inside;
+    }
+    return inside;
+}
